タイトルBGMを追加、　リザルトのドラムロールを追加 #104
