+++ Assets/Scripts/Title/TitleManager.cs
@@ -26,7 +26,7 @@ public class TitleManager : MonoBehaviour
         {
             if (!_isChangeScene)
             {
-                AudioManager.Instance.PlayVoice(14);
+                AudioManager.Instance.PlaySoundAfterExecution(Sources.VOICE, 14, () => AudioManager.Instance.PlayBGM(14));
             }
         });
     }
